+++ package.json
@@ -10,13 +10,13 @@
     "xo-fix": "xo --fix",
     "build": "npm run build-action && npm run build-win && npm run build-macos && npm run build-linux",
     "build-action": "ncc build action.js -m -s --license LICENSE",
-    "build-win": "pkg bin/cli-wrapper.js -t node12-win-x64 -o build/release/anyone-validate-fs-win",
-    "build-macos": "pkg bin/cli-wrapper.js -t node12-macos-x64 -o build/release/anyone-validate-fs-macos",
-    "build-linux": "pkg bin/cli-wrapper.js -t node12-linux-x64 -o build/release/anyone-validate-fs-linux"
+    "build-win": "pkg bin/cli.js -t node12-win-x64 -o build/release/anyone-validate-fs-win",
+    "build-macos": "pkg bin/cli.js -t node12-macos-x64 -o build/release/anyone-validate-fs-macos",
+    "build-linux": "pkg bin/cli.js -t node12-linux-x64 -o build/release/anyone-validate-fs-linux"
   },
   "bin": {
-    "anyone-validate-fs": "./bin/cli-wrapper.js",
-    "avfs": "./bin/cli-wrapper.js"
+    "anyone-validate-fs": "./bin/cli.js",
+    "avfs": "./bin/cli.js"
   },
   "repository": "git://github.com/anyone-developer/anyone-validate-fs.git",
   "keywords": [
